obituary_refset_processing/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
beir==2.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmake==4.1.0
colorama==0.4.6
datasets==3.6.0
dill==0.3.8
diskcache==5.6.3
faiss-gpu==1.7.2
filelock==3.17.0
frozenlist==1.6.0
fsspec==2025.3.0
gpt4all==2.8.2
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
llama_cpp_python==0.3.16
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pyarrow==20.0.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pytrec_eval-terrier==0.5.7
pytrials==1.0.0
pytz==2025.1
PyYAML==6.0.2
rank_bm25 @ git+https://github.com/dorianbrown/rank_bm25.git@47aa3ddf8dc1ebeb7ef4e65f2b4536af44594099
regex==2024.11.6
requests==2.32.3
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typing_extensions==4.14.1
tzdata==2025.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.20.0
